+++ src/jsonyx/_speedups.c
@@ -34,6 +34,8 @@ typedef struct _PyScannerObject {
 typedef struct _PyEncoderObject {
     PyObject_HEAD
     PyObject *Decimal;
+    PyObject *Mapping;
+    PyObject *Sequence;
     PyObject *encode_decimal;
     PyObject *indent;
     PyObject *end;
@@ -94,7 +96,7 @@ encoder_listencode_sequence(PyEncoderObject *s, PyObject *markers, _PyUnicodeWri
 static int
 encoder_listencode_obj(PyEncoderObject *s, PyObject *markers, _PyUnicodeWriter *writer, PyObject *obj, PyObject *newline_indent);
 static int
-encoder_listencode_dict(PyEncoderObject *s, PyObject *markers, _PyUnicodeWriter *writer, PyObject *dct, PyObject *newline_indent);
+encoder_listencode_mapping(PyEncoderObject *s, PyObject *markers, _PyUnicodeWriter *writer, PyObject *mapping, PyObject *newline_indent);
 static void
 raise_errmsg(const char *msg, PyObject *filename, PyObject *s, Py_ssize_t start, Py_ssize_t end);
 static PyObject *
@@ -1284,6 +1286,16 @@ encoder_new(PyTypeObject *type, PyObject *args, PyObject *kwds)
     if (s->Decimal == NULL) {
         goto bail;
     }
+    PyObject *collections_abc = PyImport_ImportModule((char *) "collections.abc");
+    if (collections_abc == NULL) {
+        goto bail;
+    }
+    s->Mapping = PyObject_GetAttrString(collections_abc, (char *) "Mapping");
+    s->Sequence = PyObject_GetAttrString(collections_abc, (char *) "Sequence");
+    Py_DECREF(collections_abc);
+    if (s->Mapping == NULL || s->Sequence == NULL) {
+        goto bail;
+    }
     s->encode_decimal = Py_NewRef(encode_decimal);
     s->indent = Py_NewRef(indent);
     s->end = Py_NewRef(end);
@@ -1428,17 +1440,20 @@ encoder_listencode_obj(PyEncoderObject *s, PyObject *markers, _PyUnicodeWriter *
             return -1;
         return _steal_accumulate(writer, encoded);
     }
-    else if (PyList_Check(obj) || PyTuple_Check(obj)) {
+    else if (PyObject_IsInstance(obj, (PyTypeObject *)s->Sequence)
+             && !PyByteArray_Check(obj) && !PyBytes_Check(obj)
+             && !PyMemoryView_Check(obj) && !PyUnicode_Check(obj))
+    {
         if (_Py_EnterRecursiveCall(" while encoding a JSON object"))
             return -1;
         rv = encoder_listencode_sequence(s, markers, writer, obj, newline_indent);
         _Py_LeaveRecursiveCall();
         return rv;
     }
-    else if (PyDict_Check(obj)) {
+    else if (PyObject_IsInstance(obj, (PyTypeObject *)s->Mapping)) {
         if (_Py_EnterRecursiveCall(" while encoding a JSON object"))
             return -1;
-        rv = encoder_listencode_dict(s, markers, writer, obj, newline_indent);
+        rv = encoder_listencode_mapping(s, markers, writer, obj, newline_indent);
         _Py_LeaveRecursiveCall();
         return rv;
     }
@@ -1514,10 +1529,10 @@ encoder_encode_key_value(PyEncoderObject *s, PyObject *markers, _PyUnicodeWriter
 }
 
 static int
-encoder_listencode_dict(PyEncoderObject *s, PyObject *markers, _PyUnicodeWriter *writer,
-                        PyObject *dct, PyObject *newline_indent)
+encoder_listencode_mapping(PyEncoderObject *s, PyObject *markers, _PyUnicodeWriter *writer,
+                           PyObject *mapping, PyObject *newline_indent)
 {
-    /* Encode Python dict dct to a JSON term */
+    /* Encode Python mapping to a JSON term */
     PyObject *ident = NULL;
     PyObject *items = NULL;
     PyObject *key, *value;
@@ -1525,11 +1540,11 @@ encoder_listencode_dict(PyEncoderObject *s, PyObject *markers, _PyUnicodeWriter 
     PyObject *new_newline_indent = NULL;
     PyObject *separator_indent = NULL;
 
-    if (PyMapping_Size(dct) == 0)  /* Fast path */
+    if (PyMapping_Size(mapping) == 0)  /* Fast path */
         return _PyUnicodeWriter_WriteASCIIString(writer, "{}", 2);
 
     int has_key;
-    ident = PyLong_FromVoidPtr(dct);
+    ident = PyLong_FromVoidPtr(mapping);
     if (ident == NULL)
         goto bail;
     has_key = PyDict_Contains(markers, ident);
@@ -1538,7 +1553,7 @@ encoder_listencode_dict(PyEncoderObject *s, PyObject *markers, _PyUnicodeWriter 
             PyErr_SetString(PyExc_ValueError, "Unexpected circular reference");
         goto bail;
     }
-    if (PyDict_SetItem(markers, ident, dct)) {
+    if (PyDict_SetItem(markers, ident, mapping)) {
         goto bail;
     }
 
@@ -1562,8 +1577,8 @@ encoder_listencode_dict(PyEncoderObject *s, PyObject *markers, _PyUnicodeWriter 
         }
     }
 
-    if (s->sort_keys || !PyDict_CheckExact(dct)) {
-        items = PyMapping_Items(dct);
+    if (s->sort_keys || !PyDict_CheckExact(mapping)) {
+        items = PyMapping_Items(mapping);
         if (items == NULL || (s->sort_keys && PyList_Sort(items) < 0))
             goto bail;
 
@@ -1586,7 +1601,7 @@ encoder_listencode_dict(PyEncoderObject *s, PyObject *markers, _PyUnicodeWriter 
 
     } else {
         Py_ssize_t pos = 0;
-        while (PyDict_Next(dct, &pos, &key, &value)) {
+        while (PyDict_Next(mapping, &pos, &key, &value)) {
             if (encoder_encode_key_value(s, markers, writer, &first, key, value,
                                          new_newline_indent,
                                          current_item_separator) < 0)