+++ docs/source/changelog.rst
@@ -21,7 +21,8 @@ v2.0.0 (unreleased)
 - Fixed comment detection
 - Fixed typo in error message
 - Improved documentation
-- Made tuples JSON serializable
+- Made :class:`collections.abc.Mapping` and :class:`collections.abc.Sequence`
+  JSON serializable
 - Removed :mod:`!jsonyx.tool`
 
 `v1.2.1 <https://pypi.org/project/jsonyx/1.2.1>`_